// src/acct-details/interface/details.interface.ts
export interface Details {
    availableBalance: string
    ledgerBalance: string
    accountNumber: string
    accountName: string
    customerId: string
    schemeType: string
    schemeCode: string
    solId: string,
    accountCurrency: string
    openEffectiveDate: Date
    accountCloseDate: Date
    status: string,
    accountType: string
    accountTier: string
    phone? : string
    email? : string
    bvn? : string
}